// interfaces/renew.ts
export type RenewListProps = {
  debt: number;
  discount_iof: string;
  discount_tic: string;
  id: number;
  installment_value: string;
  installments: number;
  loan_value: string;
  paidAmount: number;
  qtyPaid: number;
  qtyUnpaid: number;
  selected: boolean;
  tax: string;
  tax_iof: string;
  tax_tic: string;
  to_receive: string;
  unpaidAmount: number;
};